+++ packages/dash-core-components/CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- [#706](https://github.com/plotly/dash-core-components/pull/706)
+  - Added new `responsive` property that overrides the underlying Plotly.js graph responsiveness from Dash-land
+  - Added responsiveness on graph parent element resize (previously only worked on window.resize)
+  - Added new `dash-graph--pending` class to dcc.Graph, present while resizing, (re-)rendering, loading
+
+### Updated
+- [#706](https://github.com/plotly/dash-core-components/pull/706)
+  - Upgraded plotly.js to [1.51.3](https://github.com/plotly/plotly.js/releases/tag/v1.51.3)
+
 ## [1.6.0] - 2019-11-27
 ### Updated
 - Upgraded plotly.js to 1.51.2 [#708](https://github.com/plotly/dash-core-components/pull/708)